Adding user identification flag.
This enables/disables the `identifyUserWithID:andEmailAddress:` method from sending this information to the individual providers. The default is to DISABLE this functionality so previous users of this framework should be aware this will break user identification if not explicitly enabled.

This is inspired by the need to not provide user identifiable information due to the GDPR legislation. I would encourage anyone considering enabling this feature to ask the user’s consent. Please read https://techblog.bozho.net/gdpr-practical-guide-developers/ as it gives a great summary of the legislation and what is required.